requirements.txt
Django>=4.2 
djangorestframework==3.15.2 
psycopg2-binary # Adaptateur PostgreSQL pour Python, permet à Django de communiquer avec une base PostgreSQL
django-environ # Gestion simplifiée des variables d'environnement dans les projets Django
requests  # Bibliothèque HTTP simple et élégante pour faire des requêtes externes
django-cors-headers  # Gestion des en-têtes CORS (Cross-Origin Resource Sharing) dans Django
django-filter # Ajout de filtres dynamiques aux vues Django REST Framework
djangorestframework-simplejwt  # Implémentation de JWT (JSON Web Token) pour l'authentification dans DRF
python-decouple # Chargement des variables d'environnement à partir d'un fichier .env
Pillow # manipulation d'image 